fault: widen page-offset product before the multiply
In both fault-read.c and fault-write.c, a page offset into the
mapping was computed as:

    rnd_modulo_u32(total) * page_size

`page_size` is `unsigned int`, so this is a u32 * u32 product. The
multiplication is performed in 32-bit width and only promoted to the
wider destination type (size_t for pointer arithmetic on the read
side, off_t on the write side) after the multiply. At today's GB(4)
mapping cap the worst-case product is one page short of 2^32, so the
sites are safe in practice — but they are one page-size bump away
from a silent 32-bit wrap, after which the offset would land near the
base of the mapping instead of near the end.

Cast `page_size` to the destination-appropriate wide type before the
multiply so the arithmetic is performed in 64-bit width on LP64:
`(unsigned long)` for the pointer-arithmetic site, `(off_t)` for the
off_t site. This mirrors the idiom already in mlock-state.c.

Behaviour-preserving below the GB(4) cap; the change only removes the
32-bit wrap that becomes reachable if the cap is raised.

+++ mm/fault-read.c
@@ -148,7 +148,7 @@ static void read_random_pages(struct map *map)
 	for (i = 0; i < walk; i++) {
 		/* Offset is uniform across the FULL mapping; only the
 		 * iteration count is capped. */
-		char *page = p + (rnd_modulo_u32(total) * page_size);
+		char *page = p + (rnd_modulo_u32(total) * (unsigned long) page_size);
 		read_mprotect((void *) page, page_size, PROT_READ);
 		memcpy(page_buf, page, page_size);
 	}
